+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 - add protobuf format for result documents #1219 @williballenthin @mr-tz 
 
+- extractor: add Binary Ninja feature extractor @xusheng6
 - new cli flag `--os` to override auto-detected operating system for a sample @captainGeech42
 
 ### Breaking Changes